fix: Move around components to more sensible locations
