+++ app/demo/page.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import AuthenticatedLayout from "@/components/Authenticated";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import type { LitClient } from "lit-wrapper";
